Original Recipe
This is a music library organizer that I created for my girlfriend at
the time to organize disparate music libraries, where songs for a single
album could be located in multiple, randomly named, folders. The music
organizer will create a sensical folder structure, based on artist, then
album, and organize the songs accordingly.